\name{longitudinal-internal}
\alias{get.time.idx}

\title{Internal longitudinal Functions}
\description{
  Internal longitudinal functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
