\encoding{latin1}
\name{longitudinal-package}
\alias{longitudinal-package}
\docType{package}
\title{The longitudinal package}

\description{This package contains general data structures and
  functions for longitudinal data with multiple variables, 
  repeated measurements, and irregularly spaced time points.
  It also implements a shrinkage estimator of dynamical correlation
  and dynamical covariance.
}

\author{Rainer Opgen-Rhein and Korbinian Strimmer (\url{https://strimmerlab.github.io/})}
\references{
  See website: \url{https://strimmerlab.github.io/software/longitudinal/}
}
\keyword{univar}
\seealso{
\code{\link{longitudinal}, \link{dyn.cor}.}
}
